Dbrefactor invite

+++ SynchroLean/Persistence/AddUserRequestRespository.cs
@@ -21,25 +21,21 @@ namespace SynchroLean.Persistence
             await this.context.AddUserRequests.AddAsync(request);
         }
 
-        async Task<bool> IAddUserRequestRepository.AddUserRequestExists(int addUserRequestId)
+        async Task<bool> IAddUserRequestRepository.AddUserRequestExists(string email, int teamId)
         {
-            return await this.context.AddUserRequests.AnyAsync(request => request.AddUserRequestId == addUserRequestId);
+            var invite = await this.context.AddUserRequests.FindAsync(email, teamId);
+            return invite != null;
         }
 
-        async Task IAddUserRequestRepository.DeleteAddUserRequestAsync(int addUserRequestId)
+        async Task IAddUserRequestRepository.DeleteAddUserRequestAsync(string email, int teamId)
         {
-            var toDelete = await ((IAddUserRequestRepository)this).GetAddUserRequestAsync(addUserRequestId);
-            this.context.AddUserRequests.Remove(toDelete);
+            var toDelete = await this.context.AddUserRequests.FindAsync(email, teamId);
+            if(toDelete != null) this.context.AddUserRequests.Remove(toDelete);
         }
 
-        async Task<AddUserRequest> IAddUserRequestRepository.GetAddUserRequestAsync(int addUserRequestId)
+        async Task<AddUserRequest> IAddUserRequestRepository.GetAddUserRequestAsync(string email, int teamId)
         {
-            return await
-            (
-                from request in this.context.AddUserRequests
-                where request.AddUserRequestId == addUserRequestId
-                select request
-            ).SingleOrDefaultAsync();
+            return await this.context.AddUserRequests.FindAsync(email, teamId);
         }
 
         async Task<IEnumerable<AddUserRequest>> IAddUserRequestRepository.GetAddUserRequestsAsync()
